Se agregaron los servicios de patch de area, servicio y cliente. Sea gregaron los servicios de put de area, servicio y cliente. Arreglo de los parametros recibidos en Post de servicio y area. Arreglo de error del servicio delete en area.

+++ src/main/java/com/amongusdev/controller/AreaController.java
@@ -1,9 +1,13 @@
 package com.amongusdev.controller;
 
+import com.amongusdev.controller.requestdata.AreaData;
 import com.amongusdev.exception.GenericResponse;
 import com.amongusdev.models.AreaEspecializacion;
 import com.amongusdev.repositories.AreaEspecializacionRepository;
+import com.amongusdev.repositories.EspecialistaRepository;
+import com.amongusdev.repositories.ServicioRepository;
 import io.swagger.annotations.ApiOperation;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -20,6 +24,12 @@ public class AreaController {
     @Autowired
     AreaEspecializacionRepository areaEspecializacionRepository;
 
+    @Autowired
+    ServicioRepository servicioRepository;
+
+    @Autowired
+    EspecialistaRepository especialistaRepository;
+
     @GetMapping
     public List<AreaEspecializacion> listarAreas() {
         return areaEspecializacionRepository.findAll();
@@ -35,14 +45,17 @@ public class AreaController {
         }
     }
 
-    private boolean validarDatos(AreaEspecializacion area){
+    private boolean validarDatosPost(AreaEspecializacion area){
         return area.getNombre() != null;
     }
 
     @PostMapping()
-    @ApiOperation(value = "Agregar un area de especializacion", notes = "Se agrega un area de especializacion especificando con el campo nombre obligatorio")
-    public GenericResponse createArea(@Valid AreaEspecializacion area){
-        if(validarDatos(area)) {
+    @ApiOperation(value = "Agregar un area de especializacion", notes = "Se agrega un area de especializacion especificando los respectivo")
+    public GenericResponse createArea(@Valid AreaData areaData){
+        AreaEspecializacion area = new AreaEspecializacion();
+        BeanUtils.copyProperties(areaData, area);
+
+        if(validarDatosPost(area)) {
             areaEspecializacionRepository.save(area);
             return new GenericResponse(SUCCESS.getSecond(), SUCCESS.getFirst());
         } else{
@@ -54,11 +67,57 @@ public class AreaController {
     @ApiOperation(value = "Eliminar un area de especializacion", notes = "Se verifica si existe el area de especializacion y si existe se elimina")
     public GenericResponse deleteArea(@PathVariable int id){
         AreaEspecializacion area = areaEspecializacionRepository.findOne(id);
-        if(area != null) {
-            areaEspecializacionRepository.delete(id);
-            return new GenericResponse(SUCCESS.getSecond(), SUCCESS.getFirst());
+        if(area != null){
+            List<Integer> listaServicios = servicioRepository.encontrarServiciosPorArea(area.getId());
+            List<Integer> listaEspecialistas = especialistaRepository.encontrarEspecialistasPorArea(area.getId());
+            if(listaServicios.size() == 0 && listaEspecialistas.size() == 0){
+                areaEspecializacionRepository.delete(id);
+                return new GenericResponse(SUCCESS.getSecond(), SUCCESS.getFirst());
+            } else{
+                return new GenericResponse(FAILED.getSecond(), DATOS_ASOCIADOS_AREA.getSecond(), DATOS_ASOCIADOS_AREA.getFirst());
+            }
         } else{
             return new GenericResponse(FAILED.getSecond(), AREA_NOT_FOUND.getSecond(), AREA_NOT_FOUND.getFirst());
+        }
+    }
+
+    @PatchMapping("/{id}")
+    @ApiOperation(value = "Actualiza parcialmente un area de especializacion", notes = "Actualiza algunos campos especificados de un area")
+    public GenericResponse partialUpdateArea(@PathVariable int id, @Valid AreaData areaData){
+        AreaEspecializacion area = areaEspecializacionRepository.findOne(id);
+
+        if(area == null){
+            return new GenericResponse(FAILED.getSecond(), AREA_NOT_FOUND.getSecond(), AREA_NOT_FOUND.getFirst());
+        } else{
+            if(areaData.getNombre() != null)
+                area.setNombre(areaData.getNombre());
+
+            if(areaData.getDescripcion() != null)
+                area.setDescripcion(areaData.getDescripcion());
+
+            areaEspecializacionRepository.save(area);
+            return new GenericResponse(SUCCESS.getSecond(), SUCCESS.getFirst());
+        }
+    }
+
+    private boolean validarDatosPut(AreaData area){
+        return area.getNombre() != null && area.getDescripcion() != null;
+    }
+
+    @PutMapping("/{id}")
+    @ApiOperation(value = "Actualizar un area", notes = "Actualiza todos los campos de un area")
+    public GenericResponse updateArea(@PathVariable int id, @Valid AreaData areaData){
+        if(validarDatosPut(areaData)){
+            AreaEspecializacion area = areaEspecializacionRepository.findOne(id);
+            if(area == null){
+                return new GenericResponse(FAILED.getSecond(), AREA_NOT_FOUND.getSecond(), AREA_NOT_FOUND.getFirst());
+            } else {
+                BeanUtils.copyProperties(areaData, area);
+                areaEspecializacionRepository.save(area);
+                return new GenericResponse(SUCCESS.getSecond(), SUCCESS.getFirst());
+            }
+        } else{
+            return new GenericResponse(FAILED.getSecond(), FALTAN_DATOS.getSecond(), FALTAN_DATOS.getFirst());
         }
     }
 }
